Update!
renamed original file to "sprite_generator (OG)" which can be duplicated and branched off for variations.
created new version with a bunch of added functionality. the variables in the beginning of setup can now be easily changed to produce variations.

also updated the style sheet which was a broken link.